+++ src/components/TypingCodeBlock.tsx
@@ -1,18 +1,23 @@
-import { useEffect, useRef } from 'react';
-import TypingWriter from '../tools/TypingWriter';
+import { useEffect, useRef } from 'react'
+import TypingWriter from '../tools/typingWriter'
 
-import './_typingCodeBlock.scss';
-import CMDIcon from '../assets/command-window.svg';
+import './_typingCodeBlock.scss'
+import { APP_VERSION } from '../utils/constants'
+
+import CMDIcon from '..//assets/command-window.svg'
+import { useNavigate } from 'react-router-dom'
 
 export function TypingCodeBloc() {
-  const typingWriterRef = useRef<HTMLParagraphElement>(null);
-  let writer: TypingWriter | null = null;
+  const navigate = useNavigate()
+
+  const typingWriterRef = useRef<HTMLParagraphElement>(null)
+  let writer: TypingWriter | null = null
 
   useEffect(() => {
     if (typingWriterRef.current) {
-      if (!writer) writer = new TypingWriter(document.getElementById('typing-writer') as HTMLParagraphElement);
+      if (!writer) writer = new TypingWriter(document.getElementById('typing-writer') as HTMLParagraphElement)
 
-      if (writer.isActive() || writer.isFinished()) return;
+      if (writer.isActive() || writer.isFinished()) return
 
       writer
         .type('/start welcome.txt ')
@@ -32,16 +37,21 @@ export function TypingCodeBloc() {
         .pauseFor(500)
         .type('\n\n')
         .pauseFor(2000)
-        .start(true);
+        .start(true)
     }
-  }, [typingWriterRef]);
+  }, [typingWriterRef])
 
   return (
     <section className='cmd_interface font-primary'>
       <article className='cmd_interface__header select-none'>
         <div className='ml-5 w-1/2 h-full flex items-end'>
           <span className='cmd_interface__header--tab pb-2'>
-            <img src={CMDIcon} alt='Icône de console de commande' className='w-4 mr-3' />
+            <img
+              src={CMDIcon}
+              alt='Icône de console de commande'
+              className='w-4 mr-3 hover:cursor-pointer'
+              onClick={() => navigate('/dashboard')}
+            />
             @E-Code - Invite de commande
           </span>
         </div>
@@ -53,7 +63,7 @@ export function TypingCodeBloc() {
       <article className='mt-2 ml-2 flex flex-col '>
         <span>
           Site internet E-code
-          <a href='https://github.com/KinderrKill/ECode-Front/blob/main/CHANGELOG.md'> [Version 0.1]</a>
+          <a href='https://github.com/KinderrKill/ECode-Front/blob/main/CHANGELOG.md'> [Version {APP_VERSION}]</a>
         </span>
         <span>
           &copy; {new Date().getFullYear()} <a href='https://github.com/kinderrkill'>Etienne S</a>. Tout droits réservés
@@ -61,5 +71,5 @@ export function TypingCodeBloc() {
         <div ref={typingWriterRef} id='typing-writer' className='text-white mt-5 whitespace-pre-wrap text-left'></div>
       </article>
     </section>
-  );
+  )
 }