+++ src/lib/pocketBase.ts
@@ -1,0 +1,5 @@
+import PocketBase from 'pocketBase'
+
+const pocketBase = new PocketBase(import.meta.env.VITE_POCKET_BASE_URL)
+
+export default pocketBase