+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Version 0.2
+
+- Implement PocketBase (Include first test in Dashboard.tsx)
+- Add Dashboard and login page for administrators
+- Add authentification context for keeping the connexion on the website
+
 ### Version 0.1
 
 - First public release of the website